Fixed tests not running correctly with test-specific settings
